pyproject.toml

[project]
name = "user-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Yurii",email = "yurchik1997yurchik@gmail.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "litestar[standart] (>=2.15.1,<3.0.0)",
    "litestar-asyncpg (>=0.3.0,<0.4.0)",
    "advanced-alchemy (>=1.0.2,<2.0.0)",
    "msgspec (>=0.19.0,<0.20.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pydantic (>=2.11.2,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
